+++ src/app/home/page.tsx
@@ -1,0 +1,13 @@
+export default function Home() {
+  return (
+    <div>
+      <h1>Home</h1>
+      <p>Welcome to the home page of the project.</p>
+      <ul>
+        <li>Home 1</li>
+        <li>Home 2</li>
+        <li>Home 3</li>
+      </ul>
+    </div>
+  );
+}
